+++ src/overlays/actors/ovl_En_Gb2/z_en_gb2.c
@@ -698,7 +698,7 @@ void func_80B10A48(EnGb2* this, PlayState* play) {
 }
 
 void func_80B10B5C(EnGb2* this, PlayState* play) {
-    if (Player_GetMask(play) != this->unk_28C) {
+    if ((s32)Player_GetMask(play) != this->unk_28C) {
         this->unk_28C = Player_GetMask(play);
         this->unk_26C &= ~0x80;
         this->unk_26C &= ~0x20;
